Refactored out magnetic tab
