+++ src/sas/qtgui/Perspectives/Fitting/FittingLogic.py
@@ -23,6 +23,9 @@ class FittingLogic:
             self.setDataProperties()
         self.model_parameters = None
         self.kernel_module = None
+        # List of all shell-unique parameters
+        self.shell_names = []
+        self.current_shell_displayed = 0
 
     @property
     def data(self):
