+++ src/sas/qtgui/Perspectives/Fitting/FittingWidget.py
@@ -40,9 +40,9 @@ from sas.qtgui.Perspectives.Fitting import FittingUtilities
 from sas.qtgui.Perspectives.Fitting.SmearingWidget import SmearingWidget
 from sas.qtgui.Perspectives.Fitting.OptionsWidget import OptionsWidget
 from sas.qtgui.Perspectives.Fitting.PolydispersityWidget import PolydispersityWidget
+from sas.qtgui.Perspectives.Fitting.MagnetismWidget import MagnetismWidget
 from sas.qtgui.Perspectives.Fitting.FitPage import FitPage
 from sas.qtgui.Perspectives.Fitting.ViewDelegate import ModelViewDelegate
-from sas.qtgui.Perspectives.Fitting.ViewDelegate import MagnetismViewDelegate
 from sas.qtgui.Perspectives.Fitting.Constraint import Constraint
 from sas.qtgui.Perspectives.Fitting.MultiConstraint import MultiConstraint
 from sas.qtgui.Perspectives.Fitting.ReportPageLogic import ReportPageLogic
@@ -269,11 +269,6 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.model_data = None
         self._previous_model_index = 0
 
-        # Which shell is being currently displayed?
-        self.current_shell_displayed = 0
-        # List of all shell-unique parameters
-        self.shell_names = []
-
         # Error column presence in parameter display
         self.has_error_column = False
         self.has_magnet_error_column = False
@@ -321,6 +316,13 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.tabPolydispersity.setLayout(layout)
         self.lstPoly = self.polydispersity_widget.lstPoly
 
+        # magnetism widget
+        layout = QtWidgets.QGridLayout()
+        self.magnetism_widget = MagnetismWidget(parent=self, logic=self.logic)
+        layout.addWidget(self.magnetism_widget)
+        self.tabMagnetism.setLayout(layout)
+        self.lstMagnetic = self.magnetism_widget.lstMagnetic
+
         # Order widget
         layout = QtWidgets.QGridLayout()
         # pass all data items to access multiple datasets
@@ -337,13 +339,6 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.label.setText("No data loaded")
         self.lblFilename.setText("")
 
-        # Magnetic angles explained in one picture
-        self.magneticAnglesWidget = QtWidgets.QWidget()
-        labl = QtWidgets.QLabel(self.magneticAnglesWidget)
-        pixmap = QtGui.QPixmap(GuiUtils.IMAGES_DIRECTORY_LOCATION + '/M_angles_pic.png')
-        labl.setPixmap(pixmap)
-        self.magneticAnglesWidget.setFixedSize(pixmap.width(), pixmap.height())
-
     def initializeModels(self):
         """
         Set up models and views
@@ -352,8 +347,8 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         # We can't use a single model here, due to restrictions on flattening
         # the model tree with subclassed QAbstractProxyModel...
         self._model_model = FittingUtilities.ToolTippedItemModel()
-        self._poly_model = self.polydispersity_widget.polyModel()
-        self._magnet_model = FittingUtilities.ToolTippedItemModel()
+        self._poly_model = self.polydispersity_widget.poly_model
+        self._magnet_model = self.magnetism_widget._magnet_model
 
         self.model_dict["standard"] = self._model_model
         self.model_dict["poly"] = self._poly_model
@@ -364,8 +359,8 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.lst_dict["magnet"] = self.lstMagnetic
 
         self.tabToList[0] = self.lstParams
-        self.tabToList[3] = self.lstPoly
-        self.tabToList[4] = self.lstMagnetic
+        self.tabToList[3] = self.polydispersity_widget.lstPoly
+        self.tabToList[4] = self.magnetism_widget.lstMagnetic
 
         self.tabToKey[0] = "standard"
         self.tabToKey[3] = "poly"
@@ -419,11 +414,8 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.lstPoly.customContextMenuRequested.connect(self.showModelContextMenu)
 
         # Magnetism model displayed in magnetism list
-        self.lstMagnetic.setModel(self._magnet_model)
-        self.setMagneticModel()
-        FittingUtilities.setTableProperties(self.lstMagnetic)
-        # Delegates for custom editing and display
-        self.lstMagnetic.setItemDelegate(MagnetismViewDelegate(self))
+        self.magnetism_widget.setMagneticModel()
+
         # Initial status of the ordering tab - invisible
         self.tabFitting.removeTab(TAB_ORDERING)
 
@@ -580,14 +572,11 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.cmdFit.clicked.connect(self.onFit)
         self.cmdPlot.clicked.connect(self.onPlot)
         self.cmdHelp.clicked.connect(self.onHelp)
-        self.cmdMagneticDisplay.clicked.connect(self.onDisplayMagneticAngles)
 
         # Respond to change in parameters from the UI
         self._model_model.dataChanged.connect(self.onMainParamsChange)
-        self._magnet_model.dataChanged.connect(self.onMagnetModelChange)
         self.lstParams.selectionModel().selectionChanged.connect(self.onSelectionChanged)
         self.lstParams.installEventFilter(self)
-        self.lstMagnetic.installEventFilter(self)
 
         # Local signals
         self.batchFittingFinishedSignal.connect(self.batchFitComplete)
@@ -606,6 +595,8 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.polydispersity_widget.cmdFitSignal.connect(lambda: self.cmdFit.setEnabled(self.haveParamsToFit()))
         self.polydispersity_widget.updateDataSignal.connect(lambda: self.updateData())
         self.polydispersity_widget.iterateOverModelSignal.connect(lambda: self.iterateOverModel(self.updateFunctionCaption))
+        self.magnetism_widget.cmdFitSignal.connect(lambda: self.cmdFit.setEnabled(self.haveParamsToFit()))
+        self.magnetism_widget.updateDataSignal.connect(lambda: self.updateData())
 
         # Communicator signal
         self.communicate.updateModelCategoriesSignal.connect(self.onCategoriesChanged)
@@ -620,7 +611,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
 
     def eventFilter(self, obj, event):
         # Catch enter key presses when editing model params
-        if obj in [self.lstParams, self.lstPoly, self.lstMagnetic]:
+        if obj in [self.lstParams, self.polydispersity_widget.lstPoly, self.magnetism_widget.lstMagnetic]:
             if event.type() == QtCore.QEvent.KeyPress and event.key() in [QtCore.Qt.Key_Return, QtCore.Qt.Key_Enter]:
                 self.onKey(event)
                 return True
@@ -830,10 +821,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         """
         if not self.chkMagnetism.isChecked():
             return []
-        magnetic_model_params = [self._magnet_model.item(row).text()
-                            for row in range(self._magnet_model.rowCount())
-                            if self.isCheckable(row, model_key="magnet")]
-        return magnetic_model_params
+        return self.magnetism_widget.getParamNamesMagnet()
 
     def modifyViewOnRow(self, row, font=None, brush=None, model_key="standard"):
         """
@@ -1453,7 +1441,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         """
         self.main_params_to_fit = []
         self.polydispersity_widget.resetParameters()
-        self.magnet_params_to_fit = []
+        self.magnetism_widget.magnet_params_to_fit = []
 
     def onCustomModelChange(self):
         """
@@ -1615,56 +1603,6 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.cbModel.addItems(sorted([model for (model, _) in model_list if model != 'rpa']))
         self.cbModel.blockSignals(False)
 
-    def onMagnetModelChange(self, top, bottom):
-        """
-        Callback method for updating the sasmodel magnetic parameters with the GUI values
-        """
-        item = self._magnet_model.itemFromIndex(top)
-        model_column = item.column()
-        model_row = item.row()
-        name_index = self._magnet_model.index(model_row, 0)
-        parameter_name = str(self._magnet_model.data(name_index))
-
-        if model_column == 0:
-            value = item.checkState()
-            if value == QtCore.Qt.Checked:
-                self.magnet_params_to_fit.append(parameter_name)
-            else:
-                if parameter_name in self.magnet_params_to_fit:
-                    self.magnet_params_to_fit.remove(parameter_name)
-            self.cmdFit.setEnabled(self.haveParamsToFit())
-            # Update state stack
-            self.updateUndo()
-            return
-
-        # Extract changed value
-        try:
-            value = GuiUtils.toDouble(item.text())
-        except TypeError:
-            # Unparsable field
-            return
-        delegate = self.lstMagnetic.itemDelegate()
-
-        if model_column > 1:
-            if model_column == delegate.mag_min:
-                pos = 1
-            elif model_column == delegate.mag_max:
-                pos = 2
-            elif model_column == delegate.mag_unit:
-                pos = 0
-            else:
-                raise AttributeError("Wrong column in magnetism table.")
-            # min/max to be changed in self.logic.kernel_module.details[parameter_name] = ['Ang', 0.0, inf]
-            self.logic.kernel_module.details[parameter_name][pos] = value
-        else:
-            self.magnet_params[parameter_name] = value
-            #self.logic.kernel_module.setParam(parameter_name) = value
-            # Force the chart update when actual parameters changed
-            self.recalculatePlotData()
-
-        # Update state stack
-        self.updateUndo()
-
     def onHelp(self):
         """
         Show the "Fitting" section of help
@@ -1706,12 +1644,6 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         Calls parent's method for opening an HTML page
         """
         self.parent.showHelp(url)
-
-    def onDisplayMagneticAngles(self):
-        """
-        Display a simple image showing direction of magnetic angles
-        """
-        self.magneticAnglesWidget.show()
 
     def onFit(self):
         """
@@ -1927,7 +1859,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
 
         self.polydispersity_widget.updatePolyModelFromList(param_dict)
 
-        self.updateMagnetModelFromList(param_dict)
+        self.magnetism_widget.updateMagnetModelFromList(param_dict)
 
         # update charts
         self.onPlot()
@@ -2108,70 +2040,6 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.iterateOverModel(updateFittedValues)
         self.iterateOverModel(updatePolyValues)
         self._model_model.dataChanged.connect(self.onMainParamsChange)
-
-    def iterateOverMagnetModel(self, func):
-        """
-        Take func and throw it inside the magnet model row loop
-        """
-        for row_i in range(self._magnet_model.rowCount()):
-            func(row_i)
-
-    def updateMagnetModelFromList(self, param_dict):
-        """
-        Update the magnetic model with new parameters, create the errors column
-        """
-        assert isinstance(param_dict, dict)
-        if not dict:
-            return
-        if self._magnet_model.rowCount() == 0:
-            return
-
-        def updateFittedValues(row):
-            # Utility function for main model update
-            # internal so can use closure for param_dict
-            if self._magnet_model.item(row, 0) is None:
-                return
-            param_name = str(self._magnet_model.item(row, 0).text())
-            if param_name not in list(param_dict.keys()):
-                return
-            # modify the param value
-            param_repr = GuiUtils.formatNumber(param_dict[param_name][0], high=True)
-            self._magnet_model.item(row, 1).setText(param_repr)
-            self.logic.kernel_module.setParam(param_name, param_dict[param_name][0])
-            if self.has_magnet_error_column:
-                error_repr = GuiUtils.formatNumber(param_dict[param_name][1], high=True)
-                self._magnet_model.item(row, 2).setText(error_repr)
-
-        def createErrorColumn(row):
-            # Utility function for error column update
-            item = QtGui.QStandardItem()
-            def createItem(param_name):
-                if param_name in self.magnet_params_to_fit:
-                    error_repr = GuiUtils.formatNumber(param_dict[param_name][1], high=True)
-                else:
-                    error_repr = ""
-                item.setText(error_repr)
-            def curr_param():
-                return str(self._magnet_model.item(row, 0).text())
-
-            [createItem(param_name) for param_name in list(param_dict.keys()) if curr_param() == param_name]
-
-            error_column.append(item)
-
-        self.iterateOverMagnetModel(updateFittedValues)
-
-        if self.has_magnet_error_column:
-            self._magnet_model.removeColumn(2)
-
-        self.lstMagnetic.itemDelegate().addErrorColumn()
-        error_column = []
-        self.iterateOverMagnetModel(createErrorColumn)
-
-        # switch off reponse to model change
-        self._magnet_model.insertColumn(2, error_column)
-        FittingUtilities.addErrorHeadersToModel(self._magnet_model)
-
-        self.has_magnet_error_column = True
 
     def onPlot(self):
         """
@@ -2434,8 +2302,8 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
             # Add polydispersity to the model
             self.polydispersity_widget.setPolyModel()
             # Add magnetic parameters to the model
-            self.magnet_params = {}
-            self.setMagneticModel()
+            self.magnetism_widget.magnet_params = {}
+            self.magnetism_widget.setMagneticModel()
 
         # Now we claim the model has been loaded
         self.model_is_loaded = True
@@ -2509,7 +2377,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         self.addBackgroundToModel(self._model_model)
         self.undo_supported = temp_undo_state
 
-        self.shell_names = self.shellNamesList()
+        self.logic.shell_names = self.shellNamesList()
 
         # Add heading row
         FittingUtilities.addHeadingRowToModel(self._model_model, model_name)
@@ -3166,26 +3034,6 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         """
         self.lstParamHeaderSizes[index] = new_size
 
-    def setMagneticModel(self):
-        """
-        Set magnetism values on model
-        """
-        if not self.logic.model_parameters:
-            return
-        self._magnet_model.clear()
-        # default initial value
-        m0 = 0.5
-        for param in self.logic.model_parameters.call_parameters:
-            if param.type != 'magnetic': continue
-            if "M0" in param.name:
-                m0 += 0.5
-                value = m0
-            else:
-                value = param.default
-            self.addCheckedMagneticListToModel(param, value)
-
-        FittingUtilities.addHeadersToModel(self._magnet_model)
-
     def shellNamesList(self):
         """
         Returns list of names of all multi-shell parameters
@@ -3199,35 +3047,6 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
             for name in multi_names:
                 shell_names.append(name+str(i))
         return shell_names
-
-    def addCheckedMagneticListToModel(self, param, value):
-        """
-        Wrapper for model update with a subset of magnetic parameters
-        """
-        try:
-            basename, _ = param.name.rsplit('_', 1)
-        except ValueError:
-            basename = param.name
-        if basename in self.shell_names:
-            try:
-                shell_index = int(basename[-2:])
-            except ValueError:
-                shell_index = int(basename[-1:])
-
-            if shell_index > self.current_shell_displayed:
-                return
-
-        checked_list = [param.name,
-                        str(value),
-                        str(param.limits[0]),
-                        str(param.limits[1]),
-                        param.units]
-
-        self.magnet_params[param.name] = value
-
-        FittingUtilities.addCheckedListToModel(self._magnet_model, checked_list)
-        all_items = self._magnet_model.rowCount()
-        self._magnet_model.item(all_items-1,0).setData(param.name, role=QtCore.Qt.UserRole)
 
     def enableStructureFactorControl(self, structure_factor):
         """
@@ -3358,7 +3177,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
                 self.lstParams)
 
         self._num_shell_params = len(new_rows)
-        self.current_shell_displayed = index
+        self.logic.current_shell_displayed = index
 
         # Param values for existing shells were reset to default; force all changes into kernel module
         for row in new_rows:
@@ -3372,7 +3191,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         # Update relevant models
         self.polydispersity_widget.setPolyModel()
         if self.canHaveMagnetism():
-            self.setMagneticModel()
+            self.magnetism_widget.setMagneticModel()
 
     def onShowSLDProfile(self):
         """
@@ -4046,7 +3865,7 @@ class FittingWidget(QtWidgets.QWidget, Ui_FittingWidgetUI):
         widget = self.lstParams.indexWidget(index)
         if widget is not None and isinstance(widget, QtWidgets.QComboBox):
             widget.setCurrentIndex(widget.findText(str(multip)))
-        self.current_shell_displayed = multip
+        self.logic.current_shell_displayed = multip
 
     def updateFullModel(self, param_dict):
         """
